--- a/filters/ratelimit/leakybucket_test.go
+++ b/filters/ratelimit/leakybucket_test.go
@@ -32,6 +32,10 @@ func TestLeakyBucketFilterInvalidArgs(t *testing.T) {
 		{[]interface{}{"alabel", 1, "invalid period", 1, 1}},
 		{[]interface{}{"alabel", 1, "1s", "invalid capacity", 1}},
 		{[]interface{}{"alabel", 1, "1s", 1, "invalid increment"}},
+		{[]interface{}{"zero volume", 0, "1s", 1, 1}},
+		{[]interface{}{"zero period", 1, "0s", 1, 1}},
+		{[]interface{}{"zero capacity", 1, "1s", 0, 1}},
+		{[]interface{}{"zero increment", 1, "1s", 1, 0}},
 	} {
 		t.Run(fmt.Sprintf("test#%d", i), func(t *testing.T) {
 			_, err := spec.CreateFilter(test.args)