--- a/ratelimit/leakybucket.go
+++ b/ratelimit/leakybucket.go
@@ -1,0 +1,118 @@
+package ratelimit
+
+import (
+	"context"
+	_ "embed"
+	"fmt"
+	"time"
+
+	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/ext"
+
+	"github.com/zalando/skipper/metrics"
+	"github.com/zalando/skipper/net"
+)
+
+// The leaky bucket is an algorithm based on an analogy of how a bucket with a constant leak will overflow if either
+// the average rate at which water is poured in exceeds the rate at which the bucket leaks or if more water than
+// the capacity of the bucket is poured in all at once.
+// See https://en.wikipedia.org/wiki/Leaky_bucket
+type ClusterLeakyBucket struct {
+	capacity    int
+	emission    time.Duration
+	labelPrefix string
+	script      *net.RedisScript
+	ringClient  *net.RedisRingClient
+	metrics     metrics.Metrics
+	now         func() time.Time
+}
+
+const (
+	leakyBucketMetricPrefix  = "leakybucket.redis."
+	leakyBucketMetricLatency = leakyBucketMetricPrefix + "latency"
+	leakyBucketSpanName      = "redis_leakybucket"
+)
+
+// Implements leaky bucket algorithm as a Redis lua script.
+// Redis guarantees that a script is executed in an atomic way:
+// no other script or Redis command will be executed while a script is being executed.
+//
+// Possible optimization: substitute capacity and emission in script source code
+// on script creation in order not to send them over the wire on each call.
+// This way every distinct bucket configuration will get its own script.
+//
+// See https://redis.io/commands/eval
+//
+//go:embed leakybucket.lua
+var leakyBucketScript string
+
+// NewClusterLeakyBucket creates a class of leaky buckets of a given capacity and emission.
+// Emission is the reciprocal of the leak rate and equals the time to leak one unit.
+func NewClusterLeakyBucket(r *Registry, capacity int, emission time.Duration) *ClusterLeakyBucket {
+	return newClusterLeakyBucket(r.redisRing, capacity, emission, time.Now)
+}
+
+func newClusterLeakyBucket(ringClient *net.RedisRingClient, capacity int, emission time.Duration, now func() time.Time) *ClusterLeakyBucket {
+	labelPrefix := fmt.Sprintf("%d-%v-", capacity, emission)
+	return &ClusterLeakyBucket{
+		capacity, emission, labelPrefix,
+		ringClient.NewScript(leakyBucketScript), ringClient, metrics.Default, now,
+	}
+}
+
+// Add adds an increment amount to the bucket identified by the label.
+// It returns true if the amount was successfully added to the bucket or a time to wait for the next attempt.
+// It also returns any error occurred during the attempt.
+func (b *ClusterLeakyBucket) Add(ctx context.Context, label string, increment int) (added bool, retry time.Duration, err error) {
+	if increment > b.capacity {
+		// not allowed to add more than capacity and retry is not possible
+		return false, 0, nil
+	}
+
+	now := b.now()
+	span := b.startSpan(ctx)
+	defer span.Finish()
+	defer b.metrics.MeasureSince(leakyBucketMetricLatency, now)
+
+	added, retry, err = b.add(ctx, label, increment, now)
+	if err != nil {
+		ext.Error.Set(span, true)
+	}
+	return
+}
+
+func (b *ClusterLeakyBucket) add(ctx context.Context, label string, increment int, now time.Time) (added bool, retry time.Duration, err error) {
+	r, err := b.ringClient.RunScript(ctx, b.script,
+		[]string{b.getBucketId(label)},
+		b.capacity,
+		b.emission.Nanoseconds(),
+		increment,
+		now.UnixNano(),
+	)
+
+	if err == nil {
+		x := r.(int64)
+		if x >= 0 {
+			added, retry = true, 0
+		} else {
+			added, retry = false, time.Duration(-x)
+		}
+	}
+	return
+}
+
+func (b *ClusterLeakyBucket) getBucketId(label string) string {
+	return "lkb." + getHashedKey(b.labelPrefix+label)
+}
+
+func (b *ClusterLeakyBucket) startSpan(ctx context.Context) (span opentracing.Span) {
+	parent := opentracing.SpanFromContext(ctx)
+	if parent != nil {
+		span = b.ringClient.StartSpan(leakyBucketSpanName, opentracing.ChildOf(parent.Context()))
+	} else {
+		span = opentracing.NoopTracer{}.StartSpan("")
+	}
+	ext.Component.Set(span, "skipper")
+	ext.SpanKind.Set(span, "client")
+	return
+}