--- a/ratelimit/leakybucket.lua
+++ b/ratelimit/leakybucket.lua
@@ -1,10 +1,12 @@
 local bucket_id = KEYS[1]           -- bucket id
 local capacity = tonumber(ARGV[1])  -- bucket capacity in units (increment <= capacity)
-local emission = tonumber(ARGV[2])  -- time to leak one unit in ns (emission > 0)
+local emission = tonumber(ARGV[2])  -- time to leak one unit in microseconds (emission > 0)
 local increment = tonumber(ARGV[3]) -- increment in units (increment <= capacity)
-local now = tonumber(ARGV[4])       -- current time in ns (now >= 0)
+local now = tonumber(ARGV[4])       -- current time in microseconds (now >= 0)
 
 -- Redis stores the timestamp when bucket drains out.
+-- Lua uses double floating-point as a number type which can precisely represent integers only up to 2^53.
+-- The timestamp is stored in microseconds (and not nanoseconds) to keep values below 2^53.
 -- If bucket does not exist or is drained out, consider it empty now.
 local empty_at = redis.call("GET", bucket_id)
 if not empty_at then
@@ -24,7 +26,7 @@ local x = (capacity - increment) * emission - (empty_at - now)
 if x >= 0 then
     empty_at = empty_at + increment * emission
 
-    redis.call("SET", bucket_id, empty_at, "PX", math.ceil((empty_at - now) / 1e6))
+    redis.call("SET", bucket_id, empty_at, "PX", math.ceil((empty_at - now) / 1000))
 end
 
 return x