The `clusterLeakyBucket` filter implements ratelimiting _meter_ variant of [Leaky Bucket algorithm](https://en.wikipedia.org/wiki/Leaky_bucket)
and uses Redis as a storage. Leaky Bucket algorithm is also known as [Generic cell rate algorithm](https://en.wikipedia.org/wiki/Generic_cell_rate_algorithm).

Prior art
---

* https://github.com/throttled/throttled - implements CGRA algorithm,
uses optimistic locking i.e. two queries (get and set) to the storage per request
* https://github.com/brandur/redis-cell - implements GCRA as a Redis module
* https://github.com/rwz/redis-gcra - implements GCRA as Redis lua
script, uses Redis time, microsecond-only precision, floating point, divisions and rounding.
* https://github.com/go-redis/redis_rate - for go-redis v7 currently used by skipper implements
[basic rate limiter](https://redis.io/commands/incr#pattern-rate-limiter) as a counter,
since v8 uses lua script implementation from `rwz`.

This implementation uses Redis lua script and makes single Redis query per request.

Performance
---

Baseline:
```
$ bin/skipper -inline-routes='* -> clusterRatelimit("baseline-rps", 1, "10m") -> status(200) -> <shunt>' -access-log-disabled -enable-ratelimits -enable-swarm -swarm-redis-urls=localhost:6379
Requests      [total, rate, throughput]         33516, 558.58, 1.66
Duration      [total, attack, wait]             1m0s, 1m0s, 95.307ms
Latencies     [min, mean, 50, 90, 95, 99, max]  95.307ms, 179.158ms, 177.162ms, 190.552ms, 192.47ms, 209.496ms, 272.992ms
Bytes In      [total, mean]                     601488, 17.95
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           0.30%
Status Codes  [code:count]                      200:100  429:33416
Error Set:
429 Too Many Requests
```
demonstrates 558 rps throughput and a race condition (allowed requests > 1).

This implementation:
```
$ bin/skipper -inline-routes='* -> clusterLeakyBucket("clb-rps", "const", "1/10m") -> status(200) -> <shunt>' -access-log-disabled -enable-ratelimits -enable-swarm -swarm-redis-urls=localhost:6379
Requests      [total, rate, throughput]         95294, 1588.24, 0.02
Duration      [total, attack, wait]             1m0s, 1m0s, 65.821ms
Latencies     [min, mean, 50, 90, 95, 99, max]  1.248ms, 62.985ms,
64.572ms, 68.21ms, 69.023ms, 70.934ms, 133.449ms
Bytes In      [total, mean]                     0, 0.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           0.00%
Status Codes  [code:count]                      200:1  429:95293
Error Set:
429 Too Many Requests
```
demonstrates 1588 rps and no race condition as expected.

Signed-off-by: Alexander Yastrebov <alexander.yastrebov@zalando.de>