@AlexanderYastrebov 2020-11-14
f4342e1ff41edfc47a6fb630df3bdc438e211c90 updates Redis client from v7 to v8 which brings significant performance improvement

Baseline
```
$ bin/skipper -inline-routes='* -> clusterRatelimit("baseline-rps", 1, "10m") -> status(200) -> <shunt>' -access-log-disabled -enable-ratelimits -enable-swarm -swarm-redis-urls=localhost:6379
$ echo 'GET http://localhost:9090/' | vegeta attack -duration=1m -rate=0 --max-workers=100 | vegeta report
Requests      [total, rate, throughput]         721261, 12021.02, 0.05
Duration      [total, attack, wait]             1m0s, 1m0s, 3.66ms
Latencies     [min, mean, 50, 90, 95, 99, max]  435.526µs, 7.965ms, 7.585ms, 11.681ms, 13.121ms, 16.212ms, 34.447ms
Bytes In      [total, mean]                     12982644, 18.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           0.00%
Status Codes  [code:count]                      200:3  429:721258
Error Set:
429 Too Many Requests
```

This implementation:
```
$ bin/skipper -inline-routes='* -> clusterLeakyBucket("clb-rps", "const", "1/10m") -> status(200) -> <shunt>' -access-log-disabled -enable-ratelimits -enable-swarm -swarm-redis-urls=localhost:6379
$ echo 'GET http://localhost:9090/' | vegeta attack -duration=1m -rate=0 --max-workers=100 | vegeta report
Requests      [total, rate, throughput]         816894, 13615.38, 0.02
Duration      [total, attack, wait]             1m0s, 59.998s, 7.04ms
Latencies     [min, mean, 50, 90, 95, 99, max]  249.48µs, 6.503ms, 5.466ms, 12.585ms, 15.382ms, 21.437ms, 53.809ms
Bytes In      [total, mean]                     0, 0.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           0.00%
Status Codes  [code:count]                      200:1  429:816893
Error Set:
429 Too Many Requests
```